Remove TempdirTestCase
